+++ engine/src/cubos/engine/screen_picker/screen_picker.cpp
@@ -38,6 +38,11 @@ Framebuffer ScreenPicker::framebuffer()
 uint32_t ScreenPicker::at(int x, int y) const
 {
     y = mTextureSize.y - y - 1;
+    if (x >= mTextureSize.x || x < 0 || y >= mTextureSize.y || y < 0)
+    {
+        return UINT32_MAX;
+    }
+
     auto* texBuffer = new uint16_t[(std::size_t)mTextureSize.x * (std::size_t)mTextureSize.y * 2U];
 
     mIdTexture->read(texBuffer);